// src/types/enums/ERuleOptions.ts
export enum ERuleOptions {
    // ***** Rules with a 1:1 to payload *****
    EQUIFAX = 'EQUIFAX',
    SALE_TYPE = 'SALE_TYPE',
    MAX_ENQUIRIES_IN_THREE_MONTHS = 'MAX_ENQUIRIES_IN_THREE_MONTHS',
    GST_REGISTRATION_LENGTH = 'GST_REGISTRATION_LENGTH',
    ABN_REGISTRATION_LENGTH = 'ABN_REGISTRATION_LENGTH',
    FINANCE_TERM_IN_YEARS = 'FINANCE_TERM_IN_YEARS',
    PURCHASE_PRICE = 'PURCHASE_PRICE',
    LOAN_AMOUNT = 'LOAN_AMOUNT',
    MONTHS_AT_CURRENT_ADDRESS = 'MONTHS_AT_CURRENT_ADDRESS',
    MONTHS_AT_CURRENT_EMPLOYER = 'MONTHS_AT_CURRENT_EMPLOYER',
    APPLICANT_AGE = 'APPLICANT_AGE',
    COMPREHENSIVE_EQUIFAX_SCORE = 'COMPREHENSIVE_EQUIFAX_SCORE',
    ONE_EQUIFAX_SCORE = 'ONE_EQUIFAX_SCORE',
    CREDIT_HISTORY = 'CREDIT_HISTORY',
    RESIDENCY_STATUS = 'RESIDENCY_STATUS',
    EMPLOYMENT_STATUS = 'EMPLOYMENT_STATUS',
    LIVING_STATUS = 'LIVING_STATUS',
    ASSET_CONDITION = 'ASSET_CONDITION',
    FINANCE_TYPE = 'FINANCE_TYPE',
    ASSET_TYPE = 'ASSET_TYPE',
    // ***** End Rules with a 1:1 to payload *****
    // ***** Rules where things get a bit more complicated *****
    DEPOSIT_PERCENT = 'DEPOSIT_PERCENT',
    ASSET_AGE_AT_START_OF_LOAN = 'ASSET_AGE_AT_START_OF_LOAN',
    ASSET_AGE_AT_END_OF_LOAN = 'ASSET_AGE_AT_END_OF_LOAN',
    LTV = 'LTV',
    NEGATIVE_EQUITY_PERCENT = 'NEGATIVE_EQUITY_PERCENT',
    // todo
    // ***** End Rules where things get a bit more complicated *****
    // ***** Rules where things get a lot more complicated *****
    MONTHS_AT_CURRENT_EMPLOYER_BY_EMPLOYMENT_STATUS = 'MONTHS_AT_CURRENT_EMPLOYER_BY_EMPLOYMENT_STATUS',
    MAX_BALLOON_PERCENT_PER_LOAN_LENGTH_IN_YEARS = 'MAX_BALLOON_PERCENT_PER_LOAN_LENGTH_IN_YEARS',
    // ***** End Rules where things get a lot more complicated *****
}